pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for Grubflow app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "django", 
  "djangorestframework",
  "django-cors-headers",
  "django-environ",
  "psycopg2",
  "django-filter",
  "djangorestframework-simplejwt",
  "drf-yasg",
  "pillow"
]
